PMT #114945: Implement Lunr search
This commit:
- Fixes the index.json template: my first pass at generating JSON
resulted in json that was syntactically syntactially correct, but not
what I meant. Initialy I thought that objects seperated by commas would
mean an array: {...}, ..., {}
But no, arrays mean arrays. Packing the each page object into an array
was the solution, like so:
{ "pages": [ {...}, ..., {...}] }

- Implements the Lunr search: much of this code was lifted from the
Jekyll version of Polarhub. One notable change was the simplification of
building the query terms. I removed wrapping each query term with
wildcard chars because it was breaking search of whole phrases. For
example "liver" would return the liver lab activity, but "The liver is
the largest gland in the body" would not.

- Adds a blank content page to serve as a container for all this
javascript to hook on to.