PMT #114945: Add leading and trailing wildcards to search
This commit adds wildcards to provide what I think is a better user
experience. Here's what I mean: without wildcards Lunr has trouble
finding partial word matches, but it does really well with matching
strings. For example, w/o wildcards, when searching "the liver is the
largest gland in the body" the top result will be the page that starts
with that sentence.

Great, except say that your next search is for
"Hematopoiesis" but your not sure of the exact spelling. Not including
wildcards means that Lunr won't match until you get the exact word to
match. That seems like a crumby experience - most users, I presume, are
going to be searching for one or two terms at a time. Not the exact
strings from pages that they remember verbatim. Hence, adding wildcards.

This commit also tweaks some tests to match on a partial word, and
reduced the length of a long string match to match to the correct page.